Cleaned up the code
