+++ Assets/Scripts/UI/InputUIController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 45f42533e5bc1cc428eb12014f6ffd9c
+guid: 67964f393c01ef7438c2a742ca6d07ce
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
