--- Assets/Scripts/UI/UIController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 67964f393c01ef7438c2a742ca6d07ce
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
